requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.0.10
django-mongoengine==0.5.4
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
fonttools==4.39.2
kiwisolver==1.4.4
matplotlib==3.7.1
MetaTrader5==5.0.44
mongoengine==0.27.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.7
